moved the test file to the test folder
